Dto lib package structure refactoring #011
