--- common-data-lib/src/main/kotlin/com/sainnt/netshop/common/dto/RoleEnum.kt
@@ -1,5 +1,0 @@
-package com.sainnt.netshop.common.dto
-
-enum class RoleEnum {
-    ADMIN, MANAGER
-}
