+++ common-data-lib/src/main/kotlin/com/sainnt/netshop/common/dto/security/JwtAuthenticationTokenDto.kt
@@ -1,4 +1,4 @@
-package com.sainnt.netshop.common.dto
+package com.sainnt.netshop.common.dto.security
 
 data class JwtAuthenticationTokenDto (
     val token: String
