+++ common-data-lib/src/main/kotlin/com/sainnt/netshop/common/dto/security/RoleEnum.kt
@@ -1,0 +1,5 @@
+package com.sainnt.netshop.common.dto.security
+
+enum class RoleEnum {
+    ADMIN, MANAGER
+}
