+++ common-data-lib/src/main/kotlin/com/sainnt/netshop/common/dto/security/UserDto.kt
@@ -1,4 +1,4 @@
-package com.sainnt.netshop.common.dto
+package com.sainnt.netshop.common.dto.security
 
 data class UserDto(
     val id: Long,
